Begone nasty flickering!
Finally tracked it down, the problem is/was that our two-pass declaration builder(s)
confused the underlying libs and let them remove declarations in the closeContext()
method, since they where not encountered. And well, they were not encountered since
my PreDeclarationBuilder tried to skip as much as possible...

I had to revert parts of that and needed to set the setCompilingContexts() flag properly.
Also we have to make sure that we don't re-declare variables that are already there (hence
the part of the patch that checks the alias-declarations for 'global $foobar' stuff. Also,
whenever we find a variable and return early, we must not forget to setEncountered() it.

Hope this is not too hackish, but well - as far as I can see it works out pretty darn well.

Somebody owes me a pizza ;-)