Fix support for PHP 5.3's late static binding.
This is still not 100% correct, but since "static" is evaluated at runtime
we can not determine at parsing what class type it really is.
However, it's always gonna be either the class it was declared in or one
that extends the class it was declared in, so using that class as type
should provide better information than for example mixed.

BUG: 278085
REVIEW: 111598