Fix ShowFilesAfterGeneration to match generated file + drop %{dest} prefix
Using the var with the ShowFilesAfterGeneration entry is not needed
starting with KDevelop 5.1.1 (again), as the path will be automatically
prepended for all relative paths, as defined in
kdevplatform/plugins/appwizard/README.dox
(also is "dest" deprecated in favour of "PROJECTDIR")

Not using the var makes files easier to read & avoids typos
in the var name. So let default templates use recommended notation.