Add return typehint support for PHP language
Summary: This diff adds support for php return typehints (e.g. `function foo(): string`) to the parser. Nothing is done yet to use this in the TypeBuilder, that will come later.

Test Plan: The newly added test checks if the return typehint is correctly tokenized. Parsing itself is not tested, but the old tests work (so no regressions) and opening files with return typehints now works without parse errors. Tests will be added after using the return typehints in the TypeBuilder.

Reviewers: #kdevelop, pprkut, zhigalin

Subscribers: zhigalin, kdevelop-devel, #kdevelop

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D6643