Allow non-scalar constants in PHP language support
Summary: Newer PHP versions allow almost any expression in the definition of a constant, as long as it is actually constant. This is hard to test in the parser (and impossibly in the grammar definition) though, so this diff just allows any expression.

Test Plan: I edited some tests to make sure the new behaviour is allowed.

Reviewers: #kdevelop

Subscribers: kdevelop-devel, #kdevelop

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D6670