Fix chained print expressions.
This causes 10 first/follow conflicts, where I'm not sure how
to solve them or if that would even be feasible to do. Example:

  <? print 1 || print 2 || print 3; ?>

php-parser seems to handle this correctly, but conceptually it wouldn't
know if the second || belongs to the first booleanOrExpression or
is a new booleanOrExpression below the assignmentExpression of the
second print.

Reviewers: #kdevelop

Subscribers: kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D13547