Improve resolution of namespaced identifiers.
Summary:
The PST has problems matching qualified identifiers explicitely
set to global with declarations. To work around this we normalize
the namespaced identifier to be always absolute.

This means that namespace resolution inside findDeclarationImportHelper
should be correct now, but may need several calls to find the declaration
we're looking for according to the correct rules. Unfortunately, the rules
are different for classes, use declarations, functions, constants, etc.
That's why it really needs to be call-side decision on what fallbacks can
or can't be used.

BUG: 373416
FIXED-IN: 5.3.0

Reviewers: brauch

Reviewed By: brauch

Subscribers: brauch, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D14839