Add support for "array of type"
Summary:
When you use `foreach ( $a as $e )`, when `$a` is an 'iterator of x', then type of `$e` is set to 'x'.

This patch add a syntax in `@param` and `@return` to specifiy that `$a` is an 'array of x`: `@param x[]`.
This patch add support when `$a` is an 'array of x'.

{F7663472}

Reviewers: pprkut

Reviewed By: pprkut

Subscribers: pprkut, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D24921