Fix rename of a variable
Summary:
Renaming a variable doesn't work.
This is because in `KDevelop::DocumentChangeSetPrivate::generateNewText`, `rangeText(change.m_range, textLines)` return the value `$oldname` and `change.m_oldText` has the value `oldname`

This patch modify the parser to remove the "$" in the range of variable identifiers. As a result, we can now rename a variable, but as a side effect, the "$" is no more highlighted with the variable (the "$" are always blue, and do not take color of the variable)

{F7788333}

{F7788334}

{F7788335}

Test Plan:
Right-click on a variable
Select "Rename xxx..."
Choose a new name and click "Rename"

All the occurrences of the variable must be renamed

BUG: 317879
FIXED-IN: 5.5.0

Reviewers: pprkut

Reviewed By: pprkut

Subscribers: pprkut, kdevelop-devel

Tags: #kdevelop

Differential Revision: https://phabricator.kde.org/D25587